==> Notebooks/1.- Data Collection.ipynb <==
# Data Collection

## Introduction
In this notebook, we will focus on loading the initial dataset and the embedded data. The primary goal is to combine these datasets, ensuring that only the rows present in both data frames based on the 'id' and 'description' columns are retained.



## Step 1: Import Necessary Libraries
We start by importing the necessary libraries for data manipulation.

import pandas as pd

# Load the base dataset

data = pd.read_csv('../Data/DF_complete/base.csv')
print("Base Data Loaded:")
print(data.head(2))
print("Shape of the base data:", data.shape)

# Load the embedded dataset

# List of file paths
file_paths = [#'../Data/DF_embedded/df_1-embed.csv', 
              #'../Data/DF_embedded/df_2-embed.csv', 
              #'../Data/DF_embedded/df_3-embed.csv',
              #'../Data/DF_embedded/df_4-embed.csv',
              #'../Data/DF_embedded/df_5-embed.csv',
              '../Data/DF_embedded/df_6-embed.csv',
              #'../Data/DF_embedded/df_7-embed.csv',
              #'../Data/DF_embedded/df_8-embed.csv',
              #'../Data/DF_embedded/df_9-embed.csv',
              #'../Data/DF_embedded/df_10-embed.csv'
              ]

# Initialize an empty list to hold the dataframes
dataframes = []

for file in file_paths:
    # Load the dataset
    df = pd.read_csv(file)
    
    # Drop the first row
    df = df.drop(df.index[0])
    
    # Append the modified dataframe to the list
    dataframes.append(df)

# Concatenate all dataframes into one
embed_df = pd.concat(dataframes, ignore_index=True)

# Display the concatenated dataframe
print(embed_df)

Data Loading: We load two datasets – the original base.csv and the embedded data df_2-embed.csv.

## Step 2: Merging Datasets
We combine the base data and the embedded data, keeping only the rows that are present in both dataframes based on the 'id' and 'description' columns.

# Merge the datasets on 'id' and 'description'

df = data.merge(embed_df, on=['id', 'description'])
print("Merged Data:")
print(df.head(2))
print("Shape of the merged data:", df.shape)

# Save the merged data for the next notebook

df = df.sample(10000)

df.to_csv('../Data/Clean-Data/merged_data.csv', index=False)

df.shape

Merging: The merge function combines the two datasets on the common columns, ensuring only matching rows are retained.


## Conclusion
The merged dataset is saved as merged_data.csv for further processing in the next notebook.
