==> Notebooks/2.- Data Cleaning and Wrangling.ipynb <==
# Data Cleaning and Wrangling

## Introduction
In this notebook, we will clean and wrangle the merged data from the previous notebook. We will handle missing values, filter the data based on specific conditions, and prepare the dataset for exploratory data analysis (EDA).

## Step 1: Load the Merged Data
We start by loading the merged data generated from the previous notebook.

import pandas as pd
#import boxcox
from scipy.stats import boxcox
# import numpy sqrt
from numpy import sqrt, log1p
#import the functions in another file with sys.path.append
import sys
sys.path.append('../SCR/')
from functions import *

# Load the merged data
df = pd.read_csv('../Data/Clean-Data/merged_data.csv')
print("Merged Data Loaded:")
print(df.head(2))
print("Shape of the merged data:", df.shape)

Loading Data: We load the merged dataset to begin the cleaning process.

## Step 2: Handling Missing Values
We will check for and handle any missing values in the dataset.

### Check for missing values

missing_values = df.isnull().sum()
print("Missing values in each column:\n", missing_values)

### Drop rows with missing descriptions

df.dropna(subset=['description'], inplace=True)

print("Data after handling missing values:")
print(df.shape)

Handling Missing Values: Critical for ensuring the dataset's integrity before further processing.

## Step 3: Filtering Data
We'll filter the data based on specific criteria, such as the number of followers, posts, and description length.

### Filter based on followers

#df = df[(df['followers'] > 500) & (df['followers'] < 1500)]

### Filter based on the number of posts

#df = df[df['num_posts'] >= 100]

### Filter based on description length

#df = df[(df['description'].apply(len) >= 50) & (df['description'].apply(len) <= 200)]

print("Data after filtering:")
print(df.shape)

#get the unique values of the column category in the dataframe
df['category'].unique()

Filtering: Applying filters to refine the dataset to the most relevant rows.

## Step 4: Removing outliers to further statistics.
Setting the data ready for EDA.

df.columns

# For the description column, we will calculate the length of the description
df['description_length'] = df['description'].apply(len)

### Remove outliers and normalize relevant columns

df.shape

df = remove_outliers(df, 'followers')
df = remove_outliers(df, 'interactions')
df = remove_outliers(df, 'num_posts')
df = remove_outliers(df, 'description_length')

# Apply transformations

# Ensure all values are positive where needed (especially for 'followers' and 'interactions')
df['followers'] += 1
df['interactions'] += 1
df['num_posts'] += 1
df['description_length'] += 1  # Assuming description length can be zero

''' # Apply Square Root Transformation
df['followers_trans'] = np.sqrt(df['followers'])
df['interactions_trans'] = np.sqrt(df['interactions'])
df['num_posts_trans'] = np.sqrt(df['num_posts'])
df['description_length_trans'] = np.sqrt(df['description_length'])

'''

'''# Apply Box-Cox Transformation
df['followers_trans'], _ = boxcox(df['followers'])
df['interactions_trans'], _ = boxcox(df['interactions'])
df['num_posts_trans'], _ = boxcox(df['num_posts'])
df['description_length_trans'], _ = boxcox(df['description_length'])
'''

# Apply Log Transformation
df['followers_trans'] = np.log1p(df['followers'])
df['interactions_trans'] = np.log1p(df['interactions'])
df['num_posts_trans'] = np.log1p(df['num_posts'])
df['description_length_trans'] = np.log1p(df['description_length'])


'''# Apply Square Root Transformation
df['followers_sqrt'] = np.sqrt(df['followers'])
df['interactions_sqrt'] = np.sqrt(df['interactions'])
df['num_posts_sqrt'] = np.sqrt(df['num_posts'])
df['description_length_sqrt'] = np.sqrt(df['description_length'])

# Apply Log Transformation
df['followers_log'] = np.log(df['followers'])
df['interactions_log'] = np.log(df['interactions'])
df['num_posts_log'] = np.log(df['num_posts'])
df['description_length_log'] = np.log(df['description_length'])

# Apply Box-Cox Transformation
df['followers_boxcox'], _ = boxcox(df['followers'])
df['interactions_boxcox'], _ = boxcox(df['interactions'])
df['num_posts_boxcox'], _ = boxcox(df['num_posts'])
df['description_length_boxcox'], _ = boxcox(df['description_length'])'''


df.columns

'''# Output the first few rows to verify transformations
print(df[['followers_sqrt', 'followers_log', 'followers_boxcox',
          'interactions_sqrt', 'interactions_log', 'interactions_boxcox',
          'num_posts_sqrt', 'num_posts_log', 'num_posts_boxcox',
          'description_length_sqrt', 'description_length_log', 'description_length_boxcox']].head())
          '''

df.shape

# Save the cleaned data for the next notebook

df.to_csv('../Data/Clean-Data/cleaned_data.csv', index=False)

Column Management: Dropping irrelevant columns simplifies the dataset for subsequent analysis.

# Conclusion
The cleaned dataset is saved as cleaned_data.csv for the next stage of the pipeline.
